backend/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiohttp-retry==2.9.1
aiosignal==1.4.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
bcrypt==4.3.0
beautifulsoup4==4.14.2
bidict==0.23.1
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dlib-bin @ file:///C:/Users/HOUSE%20OF%20COMPUTERS/Downloads/dlib_bin-19.24.2-cp39-cp39-win_amd64.whl#sha256=3ac61a810aa963510dc6dba550750dbb5224fcb02072e2663888732f4fae4619
dnspython==2.6.1
eventlet==0.33.3
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
gast==0.4.0
git-filter-repo==2.47.0
google-auth==2.41.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.75.1
h11==0.16.0
h5py==3.14.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
imutils==0.5.4
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
keras==2.12.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.9
MarkupSafe==3.0.2
matplotlib==3.9.4
mediapipe==0.10.21
ml_dtypes==0.5.3
MouseInfo==0.1.3
multidict==6.7.0
mysql-connector-python==9.4.0
numpy==1.23.5
oauthlib==3.3.1
opencv-contrib-python==4.11.0.86
opencv-python==4.8.0.74
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
playsound==1.2.2
propcache==0.4.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.23
PyGetWindow==0.0.9
PyJWT==2.10.1
pymongo==4.7.2
PyMsgBox==2.0.1
PyMySQL==1.1.2
pyparsing==3.2.5
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.3
python-socketio==5.14.1
pytweening==1.2.0
pywhatkit==5.4
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scipy==1.13.1
sentencepiece==0.2.1
simple-websocket==1.1.0
six==1.17.0
sounddevice==0.5.2
soupsieve==2.8
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
twilio==9.2.3
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==2.3.6
wikipedia==1.4.0
wrapt==1.14.2
wsproto==1.2.0
yarl==1.22.0
zipp==3.23.0
